factor radix rank and radix rank match into separate files
