+++ rocprim/include/rocprim/block/block_radix_sort.hpp
@@ -101,8 +101,12 @@ class block_radix_sort
     static constexpr unsigned int radix_bits_per_pass = 4;
 
     using bit_key_type    = typename ::rocprim::detail::radix_key_codec<Key>::bit_key_type;
-    using block_rank_type = ::rocprim::
-        block_radix_rank<BlockSizeX, radix_bits_per_pass, true, BlockSizeY, BlockSizeZ>;
+    using block_rank_type
+        = ::rocprim::block_radix_rank<BlockSizeX,
+                                      radix_bits_per_pass,
+                                      block_radix_rank_algorithm::default_algorithm,
+                                      BlockSizeY,
+                                      BlockSizeZ>;
     using bit_keys_exchange_type = ::rocprim::block_exchange<bit_key_type, BlockSizeX, ItemsPerThread, BlockSizeY, BlockSizeZ>;
     using values_exchange_type
         = ::rocprim::block_exchange<Value, BlockSizeX, ItemsPerThread, BlockSizeY, BlockSizeZ>;
